rt-tests: bump version 0.87 => 0.89
cyclictest was updated to correct some problems with CPU affinity selection
and display of help and/or execution details.
These patches were accepted by the maintainer of the cyclictest utility.
The changes were reflected in an updated release of rt-tests - v0.89
The cyclictest changes were successfully tested on both an ARM v7 and x86_64
host environment.
The OE recipe was updated locally to incorporate the newer rt-tests release
and the resulting builds were verified for the same ARM v7 target
and for qemux86.
Patch files obsoleted by the rt-tests 0.89 release were removed from the recipe
	deleted:    meta/recipes-rt/rt-tests/files/0001-rt-tests-Allow-for-user-specified-PYLIB.patch
	deleted:    meta/recipes-rt/rt-tests/files/0002-rt-tests-Break-out-install_hwlatdetect.patch
	deleted:    meta/recipes-rt/rt-tests/files/added-missing-dependencies.patch

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>
Signed-off-by: Gary S. Robertson <gary.robertson@linaro.org>
Signed-off-by: Saul Wold <sgw@linux.intel.com>