gcc-common: Ensure checksums don't change to match old behaviour
There is a fix about to go into bitbake to ensure that datastores
being accessed with a name other than "d" are correctly reflected
in checksums. This will cause this function to add in a number of
dependencies we don't want.

These do need to be properly unravelled in due course but would
only really affect multilib builds. For now therefore just exclude
the variables as per the old behaviour.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>