sudo: upgrade from 1.8.9p5 to 1.8.10p2
Upgrade sudo from 1.8.9p5 to the newest stable release 1.8.10p2.

The license checksums are modified as required, because the doc/LICENSE
file now declares that compat/inet_pton.c bears the ISC license.

As /var/run/sudo is the default directory for sudo's time stamp
files, this patch adds a configuration file to manage this directory.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>