util-linux: fix parallel build issure for PATHFILES
When compiling util-linux, it's possible to encounter the following error.

     /bin/sh: line 2:: misc-utils/uuidd.8.tmp: No such file or directory

This is because that the misc-utils directory doesn't exist when trying to
write to misc-utils/uuidd.8.tmp.

When generating misc-utils/uuidd.8 (or anything in PATHFILES), its directory
may not have been created yet. So we need to ensure the existence of the
directory to avoid the compilation error.

[YOCTO #6292]

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>