i2c-tools: Add i2c-tools to the core
i2c-tools has been sitting outside of oe-core for long enough now. It is
a required tool for board validation, and many people are pulling it
into their builds and their own layers. Let's add it to the core.

This patch includes the i2c-tools recipe from meta-oe as of:

  commit 9df13b4140e8c6bfa0e4fb89107a6146981d2cdc
  Author: Khem Raj <raj.khem@gmail.com>
  Date:   2014-04-26

      i2c-tools: Fix build when S != B

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
Cc: Richard Purdie <richard.purdie@linuxfoundation.org>
Cc: Khem Raj <raj.khem@gmail.com>
Cc: Martin Jansa <Martin.Jansa@gmail.com>
Cc: Matthieu Crapet <Matthieu.Crapet@ingenico.com>
Cc: Koen Kooi <koen@dominion.thruhere.net>
Signed-off-by: Darren Hart <dvhart@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>