readline: exclude autoheader from autoreconf
readline ships a hand-maintained config.h, instead of letting autoheader
generate one from configure.ac.  The required arguments to AC_DEFINE are not in
configure.ac so autoheader will produce warnings and the generated code will not
behave as expected.

Solve this by excluding autoheader from autoreconf, so the upstream config.h.in
is used.

Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>