elfutils: Fix debugedit failure in grub
We've seeing occasional debugedit failures in grub during do_package
which in turn are coming from section alignment failures from libelf.
The failures occur when gold is used to link grub instead of ld.bfd.

"readelf -e uhci.module" shows:

Section Headers:
  [Nr] Name              Type            Addr     Off    Size   ES Flg Lk Inf Al
  [12] .note.GNU-stack   PROGBITS        00000000 0010ce 000000 00      0   0  1

in a good build and:

Section Headers:
  [Nr] Name              Type            Addr     Off    Size   ES Flg Lk Inf Al
  [24] .note.GNU-stack   PROGBITS        00000000 009180 000000 00      0   0  0

in a bad build. The problem is the "Al" (alignment) change from 1 to 0.
If its 0, debugedit complains.

As far as I can tell, the alignment of a zero length section is not
an issue and the check in libelf should check the section size and only
give alignment errors if there is some data to align.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>