systemd: remove udev-utils, move 'udevadm' to udev package
The udev package RDEPENDS on the udev-utils package, and the actually
udev-utils package relies on the udev package because if there's no
udev daemon, 'udevadm' command doesn't make much sense.

Now that we have removed udev-utils package and moved 'udevadm' to
the udev package in the older udev recipe, we should do the same
for systemd.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>