neard: update service file
Update the service file to more closely match the service file that has been
committed upstream.

In particular we don't want to restart neard on failure (this results in it
restarting repeatedly if no NFC hardware is found), redirecting stdout to
/dev/null means that any messages are lost instead of being sent to the journal,
and the DBus alias is required for bus activation to work correctly.

Signed-off-by: Ross Burton <ross.burton@intel.com>