toolchain-scripts: Add handling for CANADIANEXTRAOS
When we build a general toolchain script we should add all the OS
variants to PATH, not just the current one. This is because some can
cross reference each other and if the triplet prefixed ld can't be found
for example, you recieve strange errors.

Doing this resolves bugs exposed during SDK testing.

[YOCTO #6490]

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>