xmlto: add version 0.0.25
It moved from meta-oe and fixed the defect that xmlto/xsltproc
stylesheets cannot be found even when they are installed in
sysroot.

About the config files (catalog.xml), we refered what Ubuntu
13.04 did, so it could correctly search the xsl stylesheets
and xml dtds.

Assigned the environment variable XML_CATALOG_FILES, so the
xmlto will use oe-core's config as priority to avoid the
the search from build system.

[YOCTO #2416]

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>