eglibc: rewritten tzselect as posix sh and move it to tzcode package
The problem is that this would cause bash installed to meta-toolchain,
the dependencies chain is:
meta-toolchain -> eglibc-utils -> bash
We should eliminate bash dependencies in anything core to a tiny system.

The eglibc-utils also provides eglibc-utils/usr/bin/zdump,
eglibc-utils/usr/bin/zic which all of them belonged to tzcode, so we
should split these three utils as tzcode package.

[YOCTO #6544]

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>