base-files: fix up misleading dir grouping names for lsb dirs
In commit b0df35f47fb79dc149504bd66d1186b3276f9510
("base-files: Add to make some directories needed by LSB.") there
were new groupings of dirs made.  But rather than keep the grouping
names permissions based as they clearly were, it seems that it was
unfortunately misinterpreted as a counter, simply incrementing as
a new group was added.

This leads to dir3755 getting chmod'd to 0755 and dir4775 being
chmod'd to 2775 which to any new reader of the code is terribly
confusing.

Choose names that clearly reflect the permissions, and add a lsb
suffix to indicate the role.

Also note that the settings for /var/mail seem incorrect, and so
they have been aligned with what is seen in most common distros.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>