bootchart2: create recipe for bootchart2
This recipe creates packages for the bootchart2 system-wide profiler daemon
and related utilities. It fetches the Git revision immediately past the one
corresponding to the 0.14.6 release of bootchart2. (0.14.6 had a systemd-
related bug that was corrected right after it was tagged.)

The recipe contains three packages:
   *   bootchart2 - The daemon itself.
   *   pybootchartgui - Python program to visualize and display the data
       collected by bootchart2 or compatible daemons such as the original
       bootchart.
   *   bootchartd-stop-initscript - A SysV init script to stop data collection
       when booting completes.

Depending on how you wish to use bootchart2, you may not end up having all
three of those packages installed on your image.

There is also a bootchart2-native variant, which is intended solely to provide
a native version of the pybootchartgui utility. The non-cross-compiled version
of the bootchart2 daemon has not been tested at all, don't use it.

The recipe is extensively documented. Read the comments at the beginning of
bootchart2_git.bb or else you'll have no idea how to use it.

This recipe is based on a recipe from meta-WebOS. The WebOS people had some
extra code (including patches against the bootchart2 code) to support the
Upstart init system. However, since upstream Poky does not support Upstart,
that stuff is being left behind. The WebOS people can write a bbappend to re-
add it.

Original recipe written by Wonhong Kwon of LG. Upstreamed as part of the
solution to [YOCTO #5893].

Signed-off-by: Max Eliaser <max.eliaser@intel.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>